+++ src/main/java/com/anchor/domain/mentor/api/controller/MentorController.java
@@ -56,5 +56,4 @@ public class MentorController {
                          .build();
   }
 
-
 }
