+++ src/main/java/com/anchor/domain/mentor/api/controller/request/MentoringUnavailableTimeInfos.java
@@ -1,7 +1,6 @@
 package com.anchor.domain.mentor.api.controller.request;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.LocalDateTime;
+import com.anchor.global.util.DateTimeRange;
 import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,26 +10,5 @@ import lombok.NoArgsConstructor;
 public class MentoringUnavailableTimeInfos {
 
   private List<DateTimeRange> dateTimeRanges;
-
-  @Getter
-  @NoArgsConstructor
-  public static class DateTimeRange {
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "Asia/Seoul")
-    private LocalDateTime from;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "Asia/Seoul")
-    private LocalDateTime to;
-
-    private DateTimeRange(LocalDateTime from, LocalDateTime to) {
-      this.from = from;
-      this.to = to;
-    }
-
-    public static DateTimeRange of(LocalDateTime from, LocalDateTime to) {
-      return new DateTimeRange(from, to);
-    }
-
-  }
 
 }
